Moved the utility function in another package.
