+++ ocean_lib/utils/test/test_utilities.py
@@ -5,6 +5,7 @@
 import pytest
 
 from ocean_lib.utils import utilities
+from ocean_lib.utils.utilities import get_chain_id_from_url
 
 
 @pytest.mark.unit
@@ -14,3 +15,10 @@ def test_convert():
     text_bytes = utilities.convert_to_bytes(input_text)
     print("output %s" % utilities.convert_to_string(text_bytes))
     assert input_text == utilities.convert_to_text(text_bytes)
+
+
+@pytest.mark.unit
+def test_chain_id_from_url(config):
+    chain_id = get_chain_id_from_url(config["RPC_URL"])
+    assert isinstance(chain_id, int)
+    assert chain_id == 8996
